+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Replaced Axios with node-fetch
+
 ## [1.2.1] - 2021-08-03
 
 ### Added